Rename ZealChip8CompatibilityList.txt to CompatibilityList.TXT
If the ZealOS CHIP8 github repo is cloned into the ZealOS src/ folder before an ISO build or sync script execution, it will copy this file into the source tree. However, the filename is longer than the OS allows. This commit truncates it to fit comfortably within the filename string length range.
The OS will also not recognize a lowercase .txt extension as a text file, but will for capital .TXT extensions. This commit also fixes this so the file can be read natively in the OS as a text file.